% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{b_env_v}
\alias{b_env_v}
\title{make a backup of the environmental variables}
\usage{
b_env_v(v = NULL)
}
\arguments{
\item{v}{character vector with environemntal variables to be backed-up}
}
\description{
make a backup of the environmental variables
}
\keyword{internal}
