% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid.R
\name{get_cv}
\alias{get_cv}
\title{get CV}
\usage{
get_cv(
  out_file = file.path(getwd(), "CV"),
  orcid = Sys.getenv("ORCID_ID"),
  output_type = c("Rmd", "pdf"),
  json_path = NULL,
  entries = c("education", "employment", "r_package", "citation")
)
}
\arguments{
\item{out_file}{string output file}

\item{orcid}{string with the ORCID ID
by default fetched fomr "ORCID_ID" system variable}

\item{output_type}{string with output format ('Rmd' or 'pdf')}

\item{json_path}{string to JSON file with additional/custom data}

\item{entries}{character vector with entries to be included in the CV}
}
\description{
get CV as Rmd and/or PDF file with data from ORCID
}
\keyword{vitae}
