% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid.R
\name{get_personal_data}
\alias{get_personal_data}
\title{get personal data from ORCID}
\usage{
get_personal_data(orcid = Sys.getenv("ORCID_ID"))
}
\arguments{
\item{orcid}{string with the ORCID ID
by default fetched fomr "ORCID_ID" system variable}
}
\value{
list with personal data
}
\description{
get personal data from ORCID
}
\keyword{CV}
