% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid.R
\name{get_template_settings}
\alias{get_template_settings}
\title{get_template_settings}
\usage{
get_template_settings(template = "vitae::awesomecv", template_args = NULL)
}
\arguments{
\item{template}{string with the name of the CV template.
Currently only vitae templates are supported.}

\item{template_args}{list with additional arguments to be passed to the template}
}
\description{
get given template and its arguments
}
