% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitae.R
\name{get_vitae_entry}
\alias{get_vitae_entry}
\title{get vitae entry with data from ORCID or CRAN/BioC}
\usage{
get_vitae_entry(
  entry_type,
  orcid = Sys.getenv("ORCID_ID"),
  why = NULL,
  json_path = NULL
)
}
\arguments{
\item{entry_type}{string with the type of entry to be fetched}

\item{orcid}{string with the ORCID ID
by default fetched fomr "ORCID_ID" system variable}

\item{why}{list with the data for the 'why' vitae column}

\item{json_path}{string with the path to the JSON file}
}
\value{
vitae entry (standardized data.table)
}
\description{
get vitae entry with data from ORCID or CRAN/BioC
}
\keyword{vitae}
