% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{gjd}
\alias{gjd}
\title{get JSON data}
\usage{
gjd(s = NULL, names_only = FALSE, json_path = NULL)
}
\arguments{
\item{s}{string with the name of the first-level JSON entry to be fetched
if set to NULL (default) data from full JSON entry is returned}

\item{names_only}{logical, if TRUE only names
of the first-level JSON entries are returned}

\item{json_path}{string with the path to the JSON file}
}
\value{
R object returned from \code{jsonlite::fromJSON}
}
\description{
get data from json file, either subset of full data
}
