% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_env_v}
\alias{r_env_v}
\title{restore the environmental variables}
\usage{
r_env_v(rl)
}
\arguments{
\item{rl}{list with two elements: 'set' and 'unset'
Environmental variables to be set
are provided via 'set' element which is a 'Dlist'
Environmental variables to be unset
are provided via 'unset' element which is a acharacter vector}
}
\description{
restore the environmental variables
}
\keyword{internal}
