% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid.R
\name{select_vitae_template}
\alias{select_vitae_template}
\title{select vitae tepmlate}
\usage{
select_vitae_template(template = "awesomecv")
}
\arguments{
\item{template}{string with the template name}
}
\description{
select vitae template from the list of templates in package "vitae"
}
