% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rorcid.R
\name{get_cv_header}
\alias{get_cv_header}
\title{get CV header}
\usage{
get_cv_header(
  orcid = Sys.getenv("ORCID_ID"),
  json_path = NULL,
  output_type = c("list", "yaml"),
  template = "vitae::awesomecv",
  template_args = NULL
)
}
\arguments{
\item{orcid}{string with the ORCID ID
by default fetched fomr "ORCID_ID" system variable}

\item{json_path}{string to JSON file with additional/custom data}

\item{output_type}{string with output format ('list' or 'yaml'}

\item{template}{string with the name of the CV template.
Currently only vitae templates are supported.}

\item{template_args}{list with additional arguments to be passed to the template

return list or YAML with header data}
}
\description{
get CV header as YAML or list with data from ORCID
}
\keyword{CV}
